% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_modtran.R
\name{cv_modtransfer}
\alias{cv_modtransfer}
\title{Select the optimal threshold to sparsify the estimated orthogonal mapping matrix.}
\usage{
cv_modtransfer(
  ys,
  xs,
  yt,
  xt,
  family = "gaussian",
  nfold = 5,
  refine = F,
  threshold_vec = c(1e-05, 1e-04, 0.001, 0.01, 0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8,
    0.9)
)
}
\arguments{
\item{ys}{a vector for outcome in source data (if any).}

\item{xs}{the design matrix for source data.}

\item{yt}{a vector for outcome in target data (if any).}

\item{xt}{the design matrix for target data.}

\item{family}{a description of the error distribution and link function to be used in the model. The default is gaussian.}

\item{nfold}{the number of folds in cross-validation}

\item{refine}{whether we refine the estimated mapping matrix leveraging the first moment relationship before sparsifing. The default is not to refine the estimated mapping matrix.}

\item{threshold_vec}{a vector of candidate thresholds}
}
\value{
The optimal threshold that can be used to sparsify the mapping matrix.
}
\description{
This function applies cross-validation to select the optimal threshold to sparsify the estimated orthogonal mapping matrix.
}
