% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_one.R
\name{top_one}
\alias{top_one}
\title{Top one method for a vector}
\usage{
top_one(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A vector with entries as 1 or 0.
}
\description{
This function allows you to identify the largest entry of a vector, set it to 1 and set other entries to 0.
}
